Refactored gui components.
GuiGroup is now a UiComponent. Also fixed a bug where button groups that weren't set to selectOnHover would still be selected on hover.
