+++ bubolo/src/bubolo/ui/gui/SelectBox.java
@@ -21,7 +21,7 @@ import bubolo.util.Units;
  *
  * @author Christopher D. Canfield
  */
-public class SelectBox extends UiComponent implements Focusable {
+public class SelectBox extends PositionableUiComponent implements Focusable {
 	private final Args args;
 
 	private boolean hasFocus;
@@ -169,7 +169,7 @@ public class SelectBox extends UiComponent implements Focusable {
 	}
 
 	@Override
-	public int onMouseClicked(int screenX, int screenY) {
+	public ClickedObjectInfo onMouseClicked(int screenX, int screenY) {
 		if (contains(screenX, screenY)) {
 			gainFocus();
 			if (withinLeftArrow(screenX)) {
@@ -177,13 +177,13 @@ public class SelectBox extends UiComponent implements Focusable {
 			} else if (withinRightArrow(screenX)) {
 				selectNext();
 			}
-			return 0;
+			return new ClickedObjectInfo(this, 0);
 		}
-		return NoIndex;
+		return null;
 	}
 
 	@Override
-	public int onMouseMoved(int screenX, int screenY) {
+	public HoveredObjectInfo onMouseMoved(int screenX, int screenY) {
 		highlightLeftArrow = highlightRightArrow = false;
 
 		if (contains(screenX, screenY)) {
@@ -191,11 +191,11 @@ public class SelectBox extends UiComponent implements Focusable {
 			highlightLeftArrow = withinLeftArrow(screenX);
 			highlightRightArrow = withinRightArrow(screenX);
 
-			return 0;
+			return new HoveredObjectInfo(this, 0);
 		} else {
 			lostFocus();
 		}
-		return NoIndex;
+		return null;
 	}
 
 	private boolean contains(float screenX, float screenY) {
